refactor(onKeyStroke): put `onKeyDown`, `onKeyPressed`, `onKeyUp` under onKeyStroke
