+++ packages/core/index.ts
@@ -2,9 +2,6 @@ export * from './asyncComputed'
 export * from './createGlobalState'
 export * from './onClickOutside'
 export * from './onKeyStroke'
-export * from './onKeyDown'
-export * from './onKeyPressed'
-export * from './onKeyUp'
 export * from './onStartTyping'
 export * from './templateRef'
 export * from './unrefElement'
