--- packages/core/onKeyDown/index.md
@@ -1,24 +1,0 @@
----
-category: Sensors
----
-
-# onKeyDown
-Listen to the keydown event of the given key.
-
-Refer to [onKeyStroke](https://vueuse.org/onKeyStroke) for more details.
-
-## Usage
-
-```js
-import { onKeyDown } from '@vueuse/core'
-
-onKeyDown('ArrowDown', (e) => {
-  e.preventDefault()
-})
-```
-
-## Source
-
-[Source](https://github.com/vueuse/vueuse/blob/main/packages/core/onKeyDown/index.ts) • [Docs](https://github.com/vueuse/vueuse/blob/main/packages/core/onKeyDown/index.md)
-
-<!--FOOTER_ENDS-->
