--- packages/core/onKeyPressed/index.md
@@ -1,24 +1,0 @@
----
-category: Sensors
----
-
-# onKeyPressed
-Listen to the keypress event of the given key.
-
-Refer to [onKeyStroke](https://vueuse.org/onKeyStroke) for more details.
-
-## Usage
-
-```js
-import { onKeyPressed } from '@vueuse/core'
-
-onKeyPressed('ArrowDown', (e) => {
-  e.preventDefault()
-})
-```
-
-## Source
-
-[Source](https://github.com/vueuse/vueuse/blob/main/packages/core/onKeyPressed/index.ts) • [Docs](https://github.com/vueuse/vueuse/blob/main/packages/core/onKeyPressed/index.md)
-
-<!--FOOTER_ENDS-->
