--- packages/core/onKeyUp/index.md
@@ -1,24 +1,0 @@
----
-category: Sensors
----
-
-# onKeyUp
-Listen to the keyup event of the given key.
-
-Refer to [onKeyStroke](https://vueuse.org/onKeyStroke) for more details.
-
-## Usage
-
-```js
-import { onKeyUp } from '@vueuse/core'
-
-onKeyUp('ArrowDown', (e) => {
-  e.preventDefault()
-})
-```
-
-## Source
-
-[Source](https://github.com/vueuse/vueuse/blob/main/packages/core/onKeyUp/index.ts) • [Docs](https://github.com/vueuse/vueuse/blob/main/packages/core/onKeyUp/index.md)
-
-<!--FOOTER_ENDS-->
